Refactoring: encapsulate methods common for publish and build.
